add auth to dishes command handlers

+++ Restaurants.Application/Dishes/Commands/DeleteDishes/DeleteDishesCommandHandler.cs
@@ -1,7 +1,9 @@
 using MediatR;
 using Microsoft.Extensions.Logging;
+using Restaurants.Core.Domain.Constants;
 using Restaurants.Core.Domain.Entities;
 using Restaurants.Core.Domain.RepositoryContracts;
+using Restaurants.Core.Domain.ServiceContracts;
 using Restaurants.Core.Exceptions;
 
 namespace Restaurants.Application.Dishes.Commands.DeleteDishes;
@@ -9,6 +11,7 @@ namespace Restaurants.Application.Dishes.Commands.DeleteDishes;
 public class DeleteDishesCommandHandler(
     ILogger<DeleteDishesCommandHandler> logger,
     IRestaurantsRepository restaurantsRepository,
+    IRestaurantAuthorizationService authorizationService,
     IDishesRepository dishesRepository) : IRequestHandler<DeleteDishesCommand>
 {
     public async Task Handle(DeleteDishesCommand request, CancellationToken cancellationToken)
@@ -19,6 +22,9 @@ public class DeleteDishesCommandHandler(
 
         if (restaurant is null) throw new NotFoundException(nameof(Restaurant), request.RestaurantId.ToString());
 
+        if (authorizationService.Authorize(restaurant, ResourceOperation.Update)) throw new ForbidException();
+
+        
         await dishesRepository.DeleteAsync(restaurant.Dishes);
     }
 }
